refactor: 💡 remove route-action usage in accounts
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-12616
