+++ ui/admin/app/controllers/scopes/scope/auth-methods/auth-method/accounts/account/index.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeAuthMethodsAuthMethodAccountsAccountIndexController extends Controller {
+  @controller('scopes/scope/auth-methods/auth-method/accounts') accounts;
+}
