+++ ui/admin/app/controllers/scopes/scope/auth-methods/auth-method/accounts/new.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeAuthMethodsAuthMethodAccountsNewController extends Controller {
+  @controller('scopes/scope/auth-methods/auth-method/accounts') accounts;
+}
