+++ ui/admin/app/templates/account/change-password.hbs
@@ -18,8 +18,8 @@
           <page.body>
             <Form::Account::Password::ChangePassword
               @model={{@model}}
-              @submit={{route-action 'changePassword' @model}}
-              @cancel={{route-action 'cancel' @model}}
+              @submit={{fn this.changePassword @model}}
+              @cancel={{fn this.cancel @model}}
             />
           </page.body>
