+++ ui/admin/app/templates/scopes/scope/auth-methods/auth-method/accounts/account/index.hbs
@@ -18,14 +18,17 @@
   </page.navigation>
 
   <page.actions>
-    <AuthMethods::AuthMethod::Accounts::Account::Actions @model={{@model}} />
+    <AuthMethods::AuthMethod::Accounts::Account::Actions
+      @model={{@model}}
+      @delete={{this.accounts.delete}}
+    />
   </page.actions>
 
   <page.body>
     <Form::Account
       @model={{@model}}
-      @submit={{route-action 'save' @model}}
-      @cancel={{route-action 'cancel' @model}}
+      @submit={{fn this.accounts.save @model}}
+      @cancel={{fn this.accounts.cancel @model}}
     />
 
   </page.body>
