+++ ui/admin/app/templates/scopes/scope/auth-methods/auth-method/accounts/account/set-password.hbs
@@ -32,14 +32,17 @@
   </page.navigation>
 
   <page.actions>
-    <AuthMethods::AuthMethod::Accounts::Account::Actions @model={{@model}} />
+    <AuthMethods::AuthMethod::Accounts::Account::Actions
+      @model={{@model}}
+      @delete={{this.accounts.delete}}
+    />
   </page.actions>
 
   <page.body>
     <Form::Account::Password::SetPassword
       @model={{@model}}
-      @submit={{route-action 'setPassword' @model}}
-      @cancel={{route-action 'cancel' @model}}
+      @submit={{fn this.setPassword @model}}
+      @cancel={{fn this.accounts.cancel @model}}
     />
   </page.body>
