+++ ui/admin/tests/unit/controllers/scopes/scope/auth-methods/auth-method/accounts-test.js
@@ -1,0 +1,17 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'admin/tests/helpers';
+
+module(
+  'Unit | Controller | scopes/scope/auth-methods/auth-method/accounts',
+  function (hooks) {
+    setupTest(hooks);
+
+    // TODO: Replace this with your real tests.
+    test('it exists', function (assert) {
+      let controller = this.owner.lookup(
+        'controller:scopes/scope/auth-methods/auth-method/accounts',
+      );
+      assert.ok(controller);
+    });
+  },
+);
